requirements.txt
anyio==3.6.2
arrow==1.2.3
asgiref==3.6.0
attrs==22.2.0
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-debug-toolbar==4.0.0
django-picklefield==3.1
django-q==1.3.9
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.6
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2022.7.1
PyYAML==6.0
redis==3.5.3
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
stripe==5.2.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
